+++ v2/src/legacy/endpoint/index.ts
@@ -1,0 +1,24 @@
+// import { Router } from 'express';
+// import { Repository } from '../legacy/repository';
+// import { MysqlRepository } from '../legacy/repository';
+
+// import setProjectEndpoint from './project';
+// import setUserEndpoint from './user';
+// // import setUserEndpoint2 from './user2';
+
+// export async function getAPIEndpoints(db: Repository, config: { baseURL: string }): Promise<Router> {
+//   const apiRouter = Router();
+
+//   setProjectEndpoint(apiRouter, db);
+//   setUserEndpoint(apiRouter, db);
+
+//   return apiRouter;
+// }
+
+// export async function getAPIEndpoints2(db: MysqlRepository, config: { baseURL: string }): Promise<Router> {
+//   const apiRouter = Router();
+
+//   // setUserEndpoint2(apiRouter, db);
+
+//   return apiRouter;
+// }
