+++ v2/src/legacy/endpoint/user.ts
@@ -1,6 +1,6 @@
 import { Response, Router } from 'express';
-import { User } from '../entity/user';
-import { Repository } from '../repository';
+import { User } from '../interfaces/user/User';
+import { Repository } from '../legacy/repository';
 import { asyncRoute } from '../lib/endpoint';
 import util from '../lib/util';
 import statusCode from '../constants/statusCode';
